add error handling for camper, etc.
added error handling for camper, counselor and cabin.

you can now only add a camper if it can be associated with a suitable cabin. the cabin has to A: exist and B: have less that 4 current occupants.

added cabin checking for counselor. still not working as it returns an error when adding a counselor.